% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_real_data.R
\name{prepare_real_data}
\alias{prepare_real_data}
\title{SARIMA Simulations from Real World Data templates}
\usage{
prepare_real_data(n.draws = 10, n = 92, bootstrap = TRUE)
}
\arguments{
\item{n.draws}{number of draws}

\item{n}{length of the time series (in quarterts, incl 1 year foracast)}

\item{bootstrap}{should the errors in the simulated series be bootstraped?}
}
\description{
SARIMA Simulations from Real World Data templates
}
\examples{
real_data <- prepare_real_data(10)  # use 1000 to replicate paper
}
