% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ar1.R
\name{sim_ar1}
\alias{sim_ar1}
\title{AR1 Simulations and Estimations}
\usage{
sim_ar1(n.draws = 10, n = 92, out_path = path.package("tdmontecarlo",
  "out"))
}
\arguments{
\item{n.draws}{number of draws}

\item{n}{length of the time series (in quarterts, incl 1 year foracast)}

\item{out_path}{where to save the data}
}
\description{
AR1 Simulations and Estimations
}
\examples{
out_path = path.package("tdmontecarlo", "out")
# use n.draws = 1000 to replicate paper
real_data <- sim_ar1(n.draws = 10, out_path = out_path)  
}
