% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sarima.R
\name{sim_sarima}
\alias{sim_sarima}
\title{AR1 Simulations and Estimations}
\usage{
sim_sarima(n.draws = 10, n = 92, bootstrap = TRUE,
  out_path = path.package("tdmc", "out"))
}
\arguments{
\item{n.draws}{number of draws}

\item{n}{length of the time series (in quarterts, incl 1 year foracast)}

\item{bootstrap}{should the errors in the simulated series be bootstraped?
(instead of drawn from a normal)}

\item{out_path}{where to save the data}
}
\description{
AR1 Simulations and Estimations
}
\examples{
out_path = path.package("tdmc", "out")
# use n.draws = 1000 to replicate paper
real_data <- sim_sarima(n.draws = 10, out_path = out_path)  
}
